Ensure parallel docker builds always reuse common docker layers
This should actual be ensured by the docker build command, but it isn't.
If docker builds are executed one after another, common parts are taken from the cache and are properly reused.
However if started in parallel, common parts are executed twice, are not taken from the cache and hence are not reused.

Our approach works by analysing all dockerfiles to be processed and running common parts of the files as prerequisites.