Add ruby-3.1.0 to native and cross rubies
Ruby-3.1-x64 on Windows is UCRT based (instead of MSVCRT), which needs a new UCRT enabled toolchain.
This toolchain is not (yet) included into Ubuntu, so that we build proper deb packages in a second docker image called 'mingw64-ucrt'.
The deb packages are then copied from the new image and installed into a new build target called 'x64-mingw-ucrt' which is used for ruby-3.1-x64 on Windows.

Also update rake-compiler and hoe.