Use jdk 11 as maven does not work with jdk16/17
The current version of maven, which comes with ubuntu 20.04, does not work with jdk 16/17 as per https://bugs.launchpad.net/ubuntu/+source/maven/+bug/1930541

In order to have maven working in the docker container, the only solution I could come up with is to go back to jdk 11, while also maintaining ubuntu:20.04 as the base for the image. 

Another solution would be to use 21.10. but that's not LTS. Hopefully the new LTS version of ubuntu won't have this problem, but it's not there yet.