ParallelDockerBuild: Use hashed file names to shorten then
With the many images we now have, the file names of the common parts of docker images have grown to 220 characters.
This already exceeds the maximum length of file names of some file systems (ecryptfs).
Using a hash is less expressive, but usually the name of these files is just an internal thing.

Also add the tmp directory with these common-* files to "rake clean".