change `-linux` to `-linux-gnu` everywhere
Test gems built with the `-linux-gnu` suffix only on ruby 3.1 and
later, which are the ones that ship with rubygems 3.2.33 or later (the
version that first recognizes the `-gnu` suffix).

Also make sure we tag the `-linux-gnu` docker images as `-linux` for
backwards compatibility, and make sure we test that workflow.