Update manylinux2014 base images with version patched for EOL
With CentOS 7 becoming EOL as of 2024-06-30, `yum` will no longer work
in the `x86_64-linux-gnu` and `x86-linux-gnu` images shipped with
v1.5.1 and earlier.

In this commit, we update the `manylinux2014` base image for the
`x86_64-linux-gnu` and `x86-linux-gnu` images to point at a version
patched by `pypa` to use valid yum mirrors. See
https://github.com/pypa/manylinux/pull/1628 for more information.

Note that the ubuntu image version is the same from v1.5.0 and v1.5.1.